"Compare numbers"

+++ Task_19.py
@@ -1,7 +1,7 @@
 "Compare numbers"
 
-a = int(input("Enter first number: "))
-b = int(input("Enter second number: "))
+a = float(input("Enter first number: "))
+b = float(input("Enter second number: "))
 
 if a > b:
     print("a is greater")
